+++ Presentation/OpenTgResearcherConsole/Helpers/TgMenuHelperTable.cs
@@ -596,6 +596,10 @@ internal partial class TgMenuHelper
         table.AddRow(GetMarkup(TgLocale.InfoMessage(TgLocale.MenuDownloadSetIsFileNamingByMessage)),
             GetMarkup(tgDownloadSettings.SourceVm.Dto.IsFileNamingByMessage.ToString()));
 
+        // Restrict saving content
+        table.AddRow(GetMarkup(TgLocale.InfoMessage(TgLocale.MenuDownloadSetIsRestrictSavingContent)),
+            GetMarkup(tgDownloadSettings.SourceVm.Dto.IsRestrictSavingContent.ToString()));
+
         // Enabled filters
         var filters = (await BusinessLogicManager.StorageManager.FilterRepository.GetListAsync(TgEnumTableTopRecords.All, 0))
             .Items.Where(f => f.IsEnabled);