+++ Docs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increased the limit on the number of download streams for a user with a free license
 - Optimized loading of chats page in OpenTgResearcherDesktop
 - Updated documentation
+- Updated chats table in TgStorage
 ### Added
 - Added copying of messages on the chat page in OpenTgResearcherDesktop
 - Added lazy loading and data statistics to the chat page in OpenTgResearcherDesktop